The is the same as revision 1.4 except that table names are no longer capitalized.
Some of the database dumps that used this schema have two extra tables called
templinks and tempusers. These tables are left over from a database cleanup
script that is included in the log message for revision 1.4. No parts of the
WCES code or other parts of the schema rely on the presence of these two
tables.